Specify main camera with a `Resource` (#21)
Previously, the main camera was specified by adding a marker component to the camera entity. This was ergonomic for the end user, but this meant we had to deal with looking up the main camera all the time and deal with many edge cases.

Specifying it with a resource is slightly less ergonomic, but makes invalid states impossible, meaning we no longer need to deal with (or document) edge cases.

Partial implementation of Issue #20.